test: fix existing tests
